video: tegra: nvmap: Fix Cert-C error handling bug
In function nvmap_page_pool_resize casting nvmap_page_pool_get_unused_
pages() from unsigned long to unsigned int without checking its value
may result in lost or misinterpreted data. Fix this by changing data
type of curr to u64.

CID 455599
Bug 3512545

Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2748107
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/video/tegra/nvmap/nvmap_pp.c
@@ -518,7 +518,7 @@ int nvmap_page_pool_clear(void)
  */
 static void nvmap_page_pool_resize(struct nvmap_page_pool *pool, u32 size)
 {
-	u32 curr;
+	u64 curr;
 
 	rt_mutex_lock(&pool->lock);
